cleared unused carousel, added color palette
