+++ src/App.jsx
@@ -1,19 +1,23 @@
 import React from "react";
 import "./styles/App.scss";
-import { MyCarousel } from "./Components/Carousel/MyCarousel";
-// import { CodePen2 } from "./Components/Playgrounds/CodePen2";
+import { MyCarousel } from "./Components/Carousel/Carousel";
 import { CodePen } from "./Components/Playgrounds/CodePen";
 import { Footer } from "./Components/Footer";
+import { Col, Row, Container } from "react-bootstrap";
 
 function App(props) {
   console.log("props in app ", props);
   // var location = ""; //"https://codepen.io/johnsbarden/pen/BaoyJOB";
   return (
-    <div className="App">
-      <MyCarousel />
+    <Container fluid className="App">
+      <Row>
+        <Col>
+          <MyCarousel />
+        </Col>
+      </Row>
       <CodePen />
       <Footer />
-    </div>
+    </Container>
   );
 }
