+++ src/Components/Carousel/Carousel.scss
@@ -1,6 +1,8 @@
 .carousel {
-  max-height: 40rem;
+  max-height: 35rem;
   overflow: hidden;
-  max-width: 40rem;
+  max-width: 35rem;
   margin: 5rem;
+  min-height: 35rem;
+  min-width: 35rem;
 }
