+++ src/resources/defaultPlaygroundJS.jsx
@@ -4,7 +4,7 @@ const defaultJS = stripIndent`
             return (
               <div className="App">
                 <p className="app-header">This is a default file that I can modify and serve separately from the CodePen component...
-                <a href=''>Here's a link to it on GitHub (hopefully 😁)</a></p>
+                <a href='https://github.com/JohnSBarden/personal-website/blob/master/src/resources/defaultPlaygroundJS.jsx'>Here's a link to it on GitHub (hopefully 😁)</a></p>
               </div>
             );
           }
