+++ src/styles/colors.scss
@@ -1,0 +1,8 @@
+// https://coolors.co/177bb5-ffd700-faebd7-696a70-1a1d2d
+
+$dark: #1a1d2d;
+$light: #696a70;
+$fore: #faebd7;
+$accent: #ffd700;
+$accent2: #f082f0;
+$main: #177bb5;
